refactor(eval-workflows): 归位 Runtime Adapter 共享与 API 能力
