+++ src/dsh-plugin/core-adapter.ts
@@ -28,7 +28,7 @@ import {
   type ClaudeCliTrialState,
   type ClaudeResourceProjectionProfile,
 } from '../eval-workflows/runtime-adapter/adapters/claude-cli-resources.js';
-import { createSameProcessExecutorAdapter } from '../eval-workflows/runtime-adapter/adapters/same-process.js';
+import { createSameProcessExecutorAdapter } from '../eval-workflows/runtime-adapter/adapters/shared/same-process.js';
 import type { OmkBindingResourceLeaseAccess } from '../eval-workflows/runtime-adapter/resource-leases/types.js';
 import type { RuntimeBindingOf } from '../eval-workflows/runtime-adapter/types.js';
 import {
