+++ src/eval-workflows/runtime-adapter/adapters/index.ts
@@ -1,8 +1,8 @@
-export * from './anthropic-api.js';
+export * from './anthropic/api.js';
 export * from './claude-cli.js';
 export * from './claude-sdk.js';
 export * from './codex-cli.js';
 export * from './codex-sdk.js';
-export * from './custom-command.js';
-export * from './openai-api.js';
-export * from './same-process.js';
+export * from './custom/command.js';
+export * from './openai/api.js';
+export * from './shared/same-process.js';
