+++ src/eval-workflows/runtime-adapter/adapters/shared/api-protocol-core.ts
@@ -9,11 +9,11 @@ import {
   type ExecutorCapabilities,
   type JsonValue,
   type SchemaIdentity,
-} from '../../../evaluation-core/contracts/index.js';
+} from '../../../../evaluation-core/contracts/index.js';
 import {
   SOURCE_NEUTRAL_TRACE_WITHOUT_MOCKS_SCHEMA_DESCRIPTOR,
   SourceNeutralTraceWithoutMocksSchema,
-} from '../source-neutral-trace.js';
+} from '../../source-neutral-trace.js';
 
 export interface StatelessApiProtocolProfile {
   readonly providerId: string;
