+++ src/eval-workflows/runtime-adapter/evaluators/execution-assertions.ts
@@ -25,7 +25,7 @@ import {
 } from '../../assertions/deterministic.js';
 import type { ToolCallInfo } from '../../../executors/contracts/trace.js';
 import type { Assertion } from '../../../inputs/contracts/assertion.js';
-import type { SameProcessEvaluatorImplementation } from '../adapters/same-process.js';
+import type { SameProcessEvaluatorImplementation } from '../adapters/shared/same-process.js';
 import {
   SOURCE_NEUTRAL_TRACE_SCHEMA_DESCRIPTOR,
   SOURCE_NEUTRAL_TRACE_SCHEMA_VERSION,
